UPDATE eval: MetaQA

+++ KGQA/RoBERTa/main.py
@@ -319,7 +319,7 @@ def pad_x_collate_function(batch):
     print("xs: ", xs)
     print("ys: ", ys)
 
-    return xs, torch.tensor(ys)
+    return xs, torch.tensor(ys, dtype=torch.long)
 
 def perform_experiment(data_path, mode, neg_batch_size, batch_size, shuffle, num_workers, nb_epochs, embedding_dim, hidden_dim, relation_dim, gpu, use_cuda,patience, freeze, validate_every, hops, lr, entdrop, reldrop, scoredrop, l3_reg, model_name, decay, ls, load_from, outfile, do_batch_norm, que_embedding_model, valid_data_path=None, test_data_path=None):
     webqsp_checkpoint_folder = f"../../checkpoints/WebQSP/{model_name}_{que_embedding_model}_{outfile}/"
